ThreatWeaveSIEM/requirements.txt
streamlit
pandas
argparse
psutil
plotly
watchdog
requests
pyyaml
fastapi
uvicorn
psycopg[binary]
scikit-learn